requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML libraries
openai>=1.6.1
langchain>=0.1.0
langchain-openai>=0.0.2
faiss-cpu==1.7.4
numpy==1.24.3

# Database and storage
pymongo>=4.6.0
neo4j>=5.14.0
redis>=5.0.1

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Background tasks
celery>=5.3.4

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
